NeoWakeUP/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "neowakeup"
version = "0.1.0"
description = "Planetary simulation, control hub, and QAIJockey bridge runtime for drIpTECH."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Pillow>=10.0",
  "requests>=2.31",
  "numpy>=1.26",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["neowakeup*"]